Fixed attributes that were number only.
In HTML 3 you could have an attribute with a value of number:
<td colspan=3>
However, since HTML 4 all attributes are required to be quoted. That means all HTML attribute values are string by default. There were quite a few attributes in here that were just numbers. The colspan and rowspan ones just bit me when I was building a table. So I've gone through and converted them all to union types of number | string like most of them already are.